adminMember mofidy, add Member
