+++ src/main/java/com/dal/likeycakey/admin/model/service/AdminServiceImpl.java
@@ -1,12 +1,12 @@
 package com.dal.likeycakey.admin.model.service;
 
 import java.util.ArrayList;
-import  com.dal.likeycakey.admin.model.dao.AdminDao;
-import com.dal.likeycakey.admin.model.vo.Member;
-
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import  com.dal.likeycakey.admin.model.dao.AdminDao;
+import com.dal.likeycakey.admin.model.vo.AdminMember;
 
 
 @Service("adminService")
@@ -21,7 +21,7 @@ public class AdminServiceImpl implements AdminService {
 	}
 
 	@Override
-	public ArrayList<Member> selectList(int currentPage, int limit) {
+	public ArrayList<AdminMember> selectList(int currentPage, int limit) {
 		return adminDao.selectList(currentPage, limit);
 	}
