+++ src/main/java/com/dal/likeycakey/member/model/dao/MemberDao.java
@@ -1,0 +1,7 @@
+package com.dal.likeycakey.member.model.dao;
+
+import com.dal.likeycakey.member.model.vo.Member;
+
+public interface MemberDao {
+	public int insertMember(Member m) throws Exception;
+}
