+++ src/main/java/com/dal/likeycakey/member/model/service/MemberService.java
@@ -1,0 +1,7 @@
+package com.dal.likeycakey.member.model.service;
+
+import com.dal.likeycakey.member.model.vo.Member;
+
+public interface MemberService {
+	public int insertMember(Member m) throws Exception;
+}
