+++ src/main/java/com/dal/likeycakey/member/model/service/MemberServiceImpl.java
@@ -1,0 +1,21 @@
+package com.dal.likeycakey.member.model.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.dal.likeycakey.member.model.dao.MemberDao;
+import com.dal.likeycakey.member.model.vo.Member;
+
+@Service("memberService")
+public class MemberServiceImpl implements MemberService {
+	
+	@Autowired
+	private MemberDao memberDao;
+	
+	@Override
+	public int insertMember (Member m) throws Exception {
+		int result = memberDao.insertMember(m);
+		return result;
+	}
+
+}
