+++ src/main/java/com/dal/likeycakey/member/model/vo/Member.java
@@ -22,6 +22,22 @@ public class Member implements java.io.Serializable{
 	private String phone;
 	private String photo;
 	
+	public int getBlackcnt() {
+		return blackcnt;
+	}
+	public void setBlackcnt(int blackcnt) {
+		this.blackcnt = blackcnt;
+	}
+	public int getStatus() {
+		return status;
+	}
+	public void setStatus(int status) {
+		this.status = status;
+	}
+
+	private int blackcnt;
+	private int status;
+	
 	public String getId() {
 		return id;
 	}
@@ -88,7 +104,7 @@ public class Member implements java.io.Serializable{
 	}
 	
 	public Member(String id, String passwd, String name, int code, String email, String post, String addressBasic,
-			String addressDetail, String phone, String photo) {
+			String addressDetail, String phone, String photo, int blackcnt, int status) {
 		super();
 		this.id = id;
 		this.passwd = passwd;
@@ -100,6 +116,8 @@ public class Member implements java.io.Serializable{
 		this.addressDetail = addressDetail;
 		this.phone = phone;
 		this.photo = photo;
+		this.blackcnt = blackcnt;
+		this.status = status;
 	}
