+++ src/main/resources/mappers/admin-mapper.xml
@@ -4,7 +4,7 @@
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="AdminMapper">
 
-	<resultMap id="resultAdminMember" type="Member">
+	<resultMap id="resultAdminMember" type="adminMember">
 		<id property="id" column="M_ID" />
 		<result property="passwd" column="M_PW" />
 		<result property="name" column="M_NAME" />
