--- src/main/webapp/WEB-INF/db/member.sql
@@ -1,38 +1,0 @@
--- 회원 코드 설정 (M_NUM : PK)
-DROP TABLE MEMBER_CODE
-CREATE TABLE MEMBER_CODE
-(
-	M_NUM NUMBER,
-	M_TYPE VARCHAR2(12),
-	CONSTRAINT MCODE_PK_MNUM PRIMARY KEY (M_NUM)
-)
-
--- 일반회원 및 관리자 정보 (M_ID : PK)
-DROP TABLE MEMBER
-CREATE TABLE MEMBER
-(
-	M_ID VARCHAR2(50),
-	M_PW VARCHAR2(50),
-	M_NAME VARCHAR2(50),
-	M_EMAIL VARCHAR2(100),
-	M_POST VARCHAR2(10),
-	M_BASIC_ADDR VARCHAR2(150),
-	M_DETAIL_ADDR VARCHAR2(100),
-	M_CP VARCHAR2(50),
-	M_NUM NUMBER,
-	CONSTRAINT MEM_PK_MID PRIMARY KEY (M_ID),
-	CONSTRAINT MEM_FK_MNUM FOREIGN KEY (M_NUM) REFERENCES MEMBER_CODE (M_NUM)
-)
-
--- 기업회원 정보
-CREATE TABLE MEMBER_BIZ
-(
-	M_ID VARCHAR2(50),
-	BIZ_NAME VARCHAR2(50),
-	BIZ_PN VARCHAR2(20),
-	BIZ_NUM VARCHAR2(50),
-	BIZ_EMAIL VARCHAR2(50),
-	BIZ_DELIVERY NUMBER,
-	MASTER_NAME VARCHAR2(50),
-	CONSTRAINT MEMBIZ_FK_MID FOREIGN KEY (M_ID) REFERENCES MEMBER (M_ID)
-)
